Enable QSPI mode for Adafruit Feather RP2040 board
The GD25Q64C flash on the Feather board doesn't support two-byte writes
to the status register with the 01h command, instead it needs a 31h
command to write the 2nd status register byte. Besides that, it's
totally compatible with the Pico's W25Q080 flash.

Add a new PICO_FLASH_SUPPORTS_WRITE_STATUS2 config option that changes
the boot2 code to use 31h to write the status command. With this the
Feather is able to run in fast QSPI mode rather than 4x slower single
SPI mode (this difference is measurable in test code).

Also double the SPI clock speed, using a clock divider of 2 rather than
4. The Feather seems to work fine with a divider of 2 in my testing.

Much thanks to [1] for identifying what needs to change.

[1] https://github.com/raspberrypi/pico-sdk/pull/279#issuecomment-806162071