link: change default flash size from 2M to 8M
This is a total hack since the SDK doesn't have a way to change the
linker-script flash size without duplicating the entire linker script.
Ideally the linker scripts should be templated or otherwise customizable
but those are some more intrusive changes.

All I have is the Adafruit Feather RP2040 board with 8MB flash, so just
hard-code it globally as a hack.