fix -Warray-bounds warnings in GCC 12
GCC 12 got even more aggressive than 11.2 was, and it complains about
the bootrom, which is mapped at address 0 and so GCC thinks those
integer-to-pointer casts come from a NULL-derived pointer.

This can be fixed with the new (in 11.3) min-pagesize param: that flavor
of warning is only triggered for accesses in the first page, which
defaults to 4096.

pico_standard_link isn't quite the best place to be adding these flags,
but it's the most common interface library and avoids needing to
duplicate logic between pico_bootrom and pico_platform.