exclude Pioasm from the "all" target
There's now PIO headers in all builds via the cyw43 driver, and even
though the header might not get built, the Pioasm external project still
gets pulled in and built.

Set this project taget to EXCLUDE_FROM_ALL so it's not built as part of
"make all". It still gets built when needed when we'll actually compile
a pioasm-generated header.